requirements.txt
blinker==1.9.0
certifi==2022.12.7
click==8.2.1
colorama==0.4.6
distlib==0.3.6
filelock==3.11.0
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pipenv==2023.3.20
platformdirs==3.2.0
pyodbc==5.2.0
python-dotenv==1.1.1
SQLAlchemy==2.0.42
typing_extensions==4.14.1
virtualenv==20.21.0
virtualenv-clone==0.5.7
Werkzeug==3.1.3
